refactor(reactor)!: remove scheduler protocol requirement
ReactorKit no longer reschedules the pipeline. `Reactor.Scheduler`
typealias, the `scheduler` protocol requirement, its default
`MainScheduler.instance` implementation, and `.observe(on: scheduler)`
are all removed. `transform(action:)`, `mutate(_:)`,
`transform(mutation:)`, `reduce(_:_:)`, `transform(state:)`, and state
subscribers now run on whatever thread the upstream emits on.

The restored scheduler (#256) introduced subtle problems that
outweighed its convenience:
- `transform(mutation:)` merging an external stream bypassed the
  effect-boundary hop, so `reduce` could still run off-scheduler.
- `mutate(_:)` runs on the caller thread while `currentState` writes
  ran on `scheduler`, allowing races and stale reads from `mutate`.
- The synchronous `action.onNext -> reduce -> currentState` contract
  broke when the pipeline hop was introduced.

Users who need background reduction now apply `.observe(on:)` inside
`mutate(_:)` or on merged streams in `transform(mutation:)`.
Enforcement of the single-writer `reduce` invariant is a documented
contract, matching how Redux/TCA/Orbit-style frameworks handle it.

Consequences:
- `ObservedReactor.init` re-adds `.observe(on: MainScheduler.instance)`
  before the `MainActor.assumeIsolated` subscribe so SwiftUI consumers
  stay main-thread-safe regardless of where mutations emit.
- `ReactorObserving` comment now attributes the main-thread guarantee
  to `ObservedReactor.init`'s hop rather than the reactor.
- DEBUG-only `_checkActionDispatchContext` is retained (and simplified
  to warn on non-main dispatch) to keep action entry into the pipeline
  serialized — not as a scheduler surrogate.
- README `Scheduling` section replaced with a `Threading` section
  documenting the new contract: main-thread dispatch, `.observe(on:)`
  recipes, `transform(mutation:)` merge caution, `currentState` race
  warning, and UI-binding guidance for non-ObservedReactor consumers.
- `ReactorSchedulerTests` renamed to `ReactorDispatchTests`; scheduler
  assertions removed, state-stream-creation and DEBUG dispatch-check
  tests retained.

BREAKING CHANGE: `Reactor.Scheduler` typealias and `scheduler`
protocol requirement are removed. Reactors overriding `scheduler` no
longer compile. Reactors relying on automatic rescheduling (background
reduce via `SerialDispatchQueueScheduler`, or main-thread delivery via
the `MainScheduler` default) must add `.observe(on:)` explicitly.
`reactor.state` is delivered on whichever thread produced the final
mutation; UI consumers outside `ObservedReactor` must hop to main
themselves.
