+++ Sources/ReactorKitSwiftUI/ReactorObserving.swift
@@ -89,8 +89,10 @@ extension ReactorObserving: View where Content: View {
       // SwiftUI re-evaluates the surrounding body, restarting tracking.
       let _ = id
       // `MainActor.assumeIsolated` is safe here because:
-      // - ObservedReactor is @MainActor and its state setter always runs
-      //   on MainScheduler.
+      // - ObservedReactor is @MainActor and its init hops the reactor
+      //   state stream onto MainScheduler before writing to `state`,
+      //   so the setter always runs on the main thread regardless of
+      //   where the Reactor pipeline emitted from.
       // - BackportRegistrar.willSet (which fires this callback) is
       //   called from that setter.
       // - Using async dispatch would defer the @State update, causing a
